% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{condenseTaxa2}
\alias{condenseTaxa2}
\title{Condense a taxa table for a single read}
\usage{
condenseTaxa2(taxaTable, groupings = rep(1, nrow(taxaTable)))
}
\arguments{
\item{taxaTable}{a matrix or data.frame with hits on the rows and various levels of taxonomy in the columns}

\item{groupings}{a vector of groups e.g. read queries to condense taxa within}
}
\value{
a vector of length \code{ncol(taxaTable)} with NAs where the is not complete agreement
}
\description{
Take a table of taxonomic assignments from hits to a single read and condense it to a single vector with NAs where there are disagreements between the hits
}
\examples{
taxas<-matrix(c(
 'a','b','c','e',
 'a','b','d','e'
),nrow=2,byrow=TRUE)
condenseTaxa2(taxas)
}
