% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{getTaxonomy}
\alias{getTaxonomy}
\title{Get taxonomic ranks for a taxa}
\usage{
getTaxonomy(ids, taxaNodes, taxaNames, desiredTaxa = c("superkingdom",
  "phylum", "class", "order", "family", "genus", "species"),
  mc.cores = parallel::detectCores())
}
\arguments{
\item{ids}{a vector of ids to find taxonomy for}

\item{taxaNodes}{a nodes data.table from \code{\link{read.nodes}}}

\item{taxaNames}{a names data.table from \code{\link{read.names}}}

\item{desiredTaxa}{a vector of strings giving the desired taxa levels}

\item{mc.cores}{the number of cores to use when processing}
}
\value{
a data.frame with a row for each id and a column for each desiredTaxa
}
\description{
Take NCBI taxa IDs and get the corresponding taxa ranks from name and node data.tables
}
\examples{
1
}
\seealso{
\code{\link{read.nodes}}, \code{\link{read.names}}
}

