% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{makeNewick}
\alias{makeNewick}
\title{Create a Newick tree from taxonomy}
\usage{
makeNewick(taxa, naSub = "_")
}
\arguments{
\item{taxa}{a matrix with a row for each leaf of the tree and a column for each taxonomic classification e.g. the output from getTaxonomy}

\item{naSub}{a character string to substitute in place of NAs in the taxonomy}
}
\description{
Create a Newick formatted tree from a data.frame of taxonomic assignments
}
\examples{
taxa<-matrix(c('A','A','A','B','B','C','D','D','E','F','G','H'),nrow=3)
makeNewick(taxa)
}
\seealso{
\code{\link{getTaxonomy}}
}
