% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{read.accession2taxaid}
\alias{read.accession2taxaid}
\title{Read NCBI accession2taxid files}
\usage{
read.accession2taxaid(taxaFiles, sqlFile, n = 1e+06, vocal = TRUE)
}
\arguments{
\item{taxaFiles}{a string or vector of strings giving the path(s) to files to be read in}

\item{sqlFile}{a string giving the path where the output sqlite file should be saved}

\item{n}{an integer giving how many lines from an accession files to read at a time}

\item{vocal}{if TRUE output status messages}
}
\value{
TRUE if sucessful
}
\description{
Take NCBI accession2taxid files, keep only accession and taxa and save it as a sqlite database
}
\examples{
taxa<-c(
  "accession\\taccession.version\\ttaxid\\tgi",
  "Z17427\\tZ17427.1\\t3702\\t16569",
  "Z17428\\tZ17428.1\\t3702\\t16570",
  "Z17429\\tZ17429.1\\t3702\\t16571",
  "Z17430\\tZ17430.1\\t3702\\t16572"
)
inFile<-tempfile()
outFile<-tempfile()
writeLines(taxa,inFile)
read.accession2taxaid(inFile,outFile)
db<-RSQLite::dbConnect(RSQLite::SQLite(),dbname=outFile)
RSQLite::dbGetQuery(db,'SELECT * FROM accessionTaxa')
}
\references{
\url{ftp://ftp.ncbi.nih.gov/pub/taxonomy/accession2taxid}
}
\seealso{
\code{\link{read.nodes}}, \code{\link{read.names}}
}

