% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{read.names2}
\alias{read.names2}
\title{Read NCBI names file}
\usage{
read.names2(nameFile, sqlFile = "nameNode.sqlite", onlyScientific = TRUE,
  overwrite = FALSE)
}
\arguments{
\item{nameFile}{string giving the path to an NCBI name file to read from (both gzipped or uncompressed files are ok)}

\item{sqlFile}{a string giving the path where the output sqlite file should be saved}

\item{onlyScientific}{If TRUE, only store scientific names. If FALSE, synonyms and other types are included (increasing the potential for ambiguous taxonomic assignments).}

\item{overwrite}{If TRUE, delete names table in database if present and regenerate}
}
\value{
invisibly returns a string with path to sqlfile
}
\description{
Take an NCBI names file, keep only scientific names and convert it to a SQLite table
}
\examples{
namesText<-c(
  "1\\t|\\tall\\t|\\t\\t|\\tsynonym\\t|",
  "1\\t|\\troot\\t|\\t\\t|\\tscientific name\\t|",
  "2\\t|\\tBacteria\\t|\\tBacteria <prokaryotes>\\t|\\tscientific name\\t|",
  "2\\t|\\tMonera\\t|\\tMonera <Bacteria>\\t|\\tin-part\\t|",
  "2\\t|\\tProcaryotae\\t|\\tProcaryotae <Bacteria>\\t|\\tin-part\\t|"
)
read.names(textConnection(namesText))
}
\references{
\url{ftp://ftp.ncbi.nih.gov/pub/taxonomy/}
}
\seealso{
\code{\link{read.nodes}}
}
