% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{trimTaxa}
\alias{trimTaxa}
\title{Trim columns from taxa file}
\usage{
trimTaxa(inFile, outFile)
}
\arguments{
\item{inFile}{a single string giving the 4 column tab separated file to read from}

\item{outFile}{a single string giving the file path to write to}
}
\description{
A simple script to delete the first row and then delete the first and fourth column of a four column tab delimited file and write to another file.
}
